% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_FU.R
\name{calculate_FU}
\alias{calculate_FU}
\title{calculate_FU}
\usage{
calculate_FU(data)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date and event.}
}
\value{
data.frame object
}
\description{
help function
}
\details{
The implementation of the methods is done as follows. For each individual
i the time of entry into the study t^{start}_{i}, the final recorded
date t^{final}_{i} and a status indicator s_i (taking the value 1 if
t^{final}_{i} is the date of the event and 0 if t^{final}_{i} is the
date of censoring) is observed. Then, letting y_i = t_i^{final} - t_i^{start},
the three commonly reported measures can be all calculated using the standard
Kaplan-Meier estimator, with the following adjustments:
\itemize{
   \item time to censoring: this is calculated via a Kaplan-Meier estimate based on
\{(y_i, 1- s_i), i=1, ..., n\},
   \item observation time: this is calculated via a Kaplan-Meier estimate based
on \{(y_i, 1), i=1, ..., n\},
   \item observation time for those event-free: this is calculated via a Kaplan-Meier
estimate based on \{(y_i, 1), i = 1,..., n such that s_i=0\}.
}
}
\examples{
data <- data.frame(start_date = as.Date(rep(0, 10), origin = "2022-01-01"), final_date = as.Date(c(120, 300, 500, 750, 290, 98, 167, 250, 734, 230), origin = "2022-01-01"), event = c(1, 1, 0, 1, 0, 1, 1, 0, 1, 1))
calculate_FU(data)

}
\references{
Erdmann, S. & Betensky, R. A. (2023). KMstability: R tools to report
the stability and precision of Kaplan-Meier estimates as well as measures
of follow-up in time-to-event studies. SoftwareX (in revision).

Betensky, R. A. (2015). Measures of follow-up in time-to-event studies:
Why provide them and what should they be?. Clinical Trials, 12(4), 403-408.
}
