% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_FU_group.R
\name{calculate_FU_group}
\alias{calculate_FU_group}
\title{calculate_FU_group}
\usage{
calculate_FU_group(data)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date, event and group.}
}
\value{
data.frame object
}
\description{
help function
}
\examples{
data <- data.frame(start_date = c(as.Date(rep(0, 200), origin = "2022-01-01"),
as.Date(rep(0, 200), origin = "2022-01-01")),
final_date = c(as.Date(round(runif(200, 0, 250)), origin = "2022-01-01"),
               as.Date(round(runif(200, 0, 450)), origin = "2022-01-01")),
event = c(rbinom(200, 1, 0.9), rbinom(200, 1, 0.9)),
group = c(rep(0, 200), rep(1, 200)))
calculate_FU_group(data)

}
