% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure2.R
\name{Figure2}
\alias{Figure2}
\title{Figure2}
\usage{
Figure2(data)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date and event.}
}
\value{
ggsurvplot object
}
\description{
detailed graphical presentation of the alternative measure proposed by Betensky (2015)
}
\examples{
data <- data.frame(start_date = as.Date(rep(0, 100), origin = "2022-01-01"),
final_date = as.Date(round(runif(100, 0, 250)), origin = "2022-01-01"),
event = rbinom(100, 1, 0.9))
Figure2(data)

}
