% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure3.R
\name{Figure3}
\alias{Figure3}
\title{Figure3}
\usage{
Figure3(data)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date and event.}
}
\value{
ggsurvplot object
}
\description{
detailed graphical presentation of common measures of follow-up (i.e. time to censoring, observation time and observation time for those event-free)
}
\examples{
data <- data.frame(start_date = as.Date(rep(0, 100), origin = "2022-01-01"),
final_date = as.Date(round(runif(100, 0, 250)), origin = "2022-01-01"),
event = rbinom(100, 1, 0.9))
Figure3(data)

}
