% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure4.R
\name{Figure4}
\alias{Figure4}
\title{Figure4}
\usage{
Figure4(data)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date and event.}
}
\value{
ggsurvplot object
}
\description{
detailed graphical presentations of derived metrics of the alternative measure proposed by Betensky (2015)
}
\examples{
data <- data.frame(start_date = as.Date(rep(0, 100), origin = "2022-01-01"),
final_date = as.Date(round(runif(100, 0, 250)), origin = "2022-01-01"),
event = rbinom(100, 1, 0.9))
Figure4(data)

}
