% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FU_plot.R
\name{FU_plot}
\alias{FU_plot}
\title{FU_plot}
\usage{
FU_plot(data)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date and event.}
}
\value{
ggsurvplot object
}
\description{
detailed graphical presentation of common measures of follow-up (i.e. time to censoring, observation time and observation time for those event-free)
}
\details{
The implementation of the methods is done as follows. For each individual
i the time of entry into the study t^{start}_{i}, the final recorded
date t^{final}_{i} and a status indicator s_i (taking the value 1 if
t^{final}_{i} is the date of the event and 0 if t^{final}_{i} is the
date of censoring) is observed. Then, letting y_i = t_i^{final} - t_i^{start},
the three commonly reported measures can be all calculated using the standard
Kaplan-Meier estimator, with the following adjustments:
\itemize{
   \item time to censoring: this is calculated via a Kaplan-Meier estimate based on
\{(y_i, 1- s_i), i=1, ..., n\},
   \item observation time: this is calculated via a Kaplan-Meier estimate based
on \{(y_i, 1), i=1, ..., n\},
   \item observation time for those event-free: this is calculated via a Kaplan-Meier
estimate based on \{(y_i, 1), i = 1,..., n such that s_i=0\}.
}
}
\examples{
data <- data.frame(start_date = as.Date(rep(0, 100), origin = "2022-01-01"),
final_date = as.Date(round(runif(100, 0, 250)), origin = "2022-01-01"),
event = rbinom(100, 1, 0.9))
FU_plot(data)

}
\references{
Erdmann, S. & Betensky, R. A. (2023). KMstability: R tools to report
the stability and precision of Kaplan-Meier estimates as well as measures
of follow-up in time-to-event studies. SoftwareX (in revision).

Betensky, R. A. (2015). Measures of follow-up in time-to-event studies:
Why provide them and what should they be?. Clinical Trials, 12(4), 403-408.
}
