% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FU_tab.R
\name{FU_tab}
\alias{FU_tab}
\title{FU_tab}
\usage{
FU_tab(data, pretty = TRUE)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date and event.}
}
\value{
data.frame object
}
\description{
derived metrics of common measures of follow-up (i.e. time to censoring, observation time and observation time for those event-free)
}
\examples{
data <- data.frame(start_date = as.Date(rep(0, 100), origin = "2022-01-01"),
final_date = as.Date(round(runif(100, 0, 250)), origin = "2022-01-01"),
event = rbinom(100, 1, 0.9))
FU_tab(data)

}
