% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FU_tab_group.R
\name{FU_tab_group}
\alias{FU_tab_group}
\title{FU_tab_group}
\usage{
FU_tab_group(data)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date, event and group..}
}
\value{
data.frame object
}
\description{
derived metrics of common measures of follow-up (i.e. time to censoring, observation time and observation time for those event-free) for group comparison
}
\examples{
data <- data.frame(start_date = c(as.Date(rep(0, 200), origin = "2022-01-01"),
as.Date(rep(0, 200), origin = "2022-01-01")),
final_date = c(as.Date(round(runif(200, 0, 250)), origin = "2022-01-01"),
               as.Date(round(runif(200, 0, 450)), origin = "2022-01-01")),
event = c(rbinom(200, 1, 0.9), rbinom(200, 1, 0.9)),
group = c(rep(0, 200), rep(1, 200)))
FU_tab_group(data)

}
