% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_plot_group.R
\name{KM_plot_group}
\alias{KM_plot_group}
\title{KM_plot_group}
\usage{
KM_plot_group(data)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date, event and group.}
}
\value{
ggsurvplot object
}
\description{
detailed graphical presentations of the entire distribution via adequate Kaplan-Meier plots including confidence intervals and detailed life tables for group comparison
}
\examples{
data <- data.frame(start_date = c(as.Date(rep(0, 200), origin = "2022-01-01"),
as.Date(rep(0, 200), origin = "2022-01-01")),
final_date = c(as.Date(round(runif(200, 0, 250)), origin = "2022-01-01"),
               as.Date(round(runif(200, 0, 450)), origin = "2022-01-01")),
event = c(rbinom(200, 1, 0.9), rbinom(200, 1, 0.9)),
group = c(rep(0, 200), rep(1, 200)))
KM_plot_group(data)

}
