% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_SI_tab.R
\name{KM_SI_tab}
\alias{KM_SI_tab}
\title{KM_SI_tab}
\usage{
KM_SI_tab(data, pretty = TRUE)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date and event.}
}
\value{
data.frame object
}
\description{
derived metrics of the Kaplan-Meier estimate and alternative measure of the stability of the Kaplan-Meier estimate (Betensky, 2015)
}
\examples{
data <- data.frame(start_date = as.Date(rep(0, 100), origin = "2022-01-01"),
final_date = as.Date(round(runif(100, 0, 250)), origin = "2022-01-01"),
event = rbinom(100, 1, 0.9))
KM_SI_tab(data)

}
