% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_SI_tab_group.R
\name{KM_SI_tab_group}
\alias{KM_SI_tab_group}
\title{KM_SI_tab_group}
\usage{
KM_SI_tab_group(data)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date, event and group.}
}
\value{
data.frame object
}
\description{
derived metrics of the Kaplan-Meier estimate and alternative measure of the stability of the Kaplan-Meier estimate (Betensky, 2015)
}
\examples{
data <- data.frame(start_date = c(as.Date(rep(0, 200), origin = "2022-01-01"),
as.Date(rep(0, 200), origin = "2022-01-01")),
final_date = c(as.Date(round(runif(200, 0, 250)), origin = "2022-01-01"),
               as.Date(round(runif(200, 0, 450)), origin = "2022-01-01")),
event = c(rbinom(200, 1, 0.9), rbinom(200, 1, 0.9)),
group = c(rep(0, 200), rep(1, 200)))
KM_SI_tab_group(data)

}
