% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_FU.R
\name{calculate_FU}
\alias{calculate_FU}
\title{calculate_FU}
\usage{
calculate_FU(data)
}
\arguments{
\item{data}{data.frame containing variables named start_date, final_date and event.}
}
\value{
data.frame object
}
\description{
help function
}
\examples{
data <- data.frame(start_date = as.Date(rep(0, 10), origin = "2022-01-01"), final_date = as.Date(c(120, 300, 500, 750, 290, 98, 167, 250, 734, 230), origin = "2022-01-01"), event = c(1, 1, 0, 1, 0, 1, 1, 0, 1, 1))
calculate_FU(data)

}
